requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.5
django-crispy-forms==2.4
django-extensions==4.1
django-tinymce==4.1.0
djangorestframework==3.16.1
pillow==11.3.0
psycopg2==2.9.10
pycurl==7.45.6
sqlparse==0.5.3
tzdata==2025.2
